# openapi.yaml
{ "openapi": "3.0.1","info": { "title": "Employees service API","description": "Employees Information","version": "1.0" },"servers": [ { "url": "http://localhost:8080","description": "local server","variables": { } } ],"paths": { "/api/employees/{id}": { "get": { "tags": [ "employee-controller" ],"summary": "Get an employee by id","operationId": "getEmployee","parameters": [ { "name": "id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "500": { "description": "Internal server error","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } },"200": { "description": "Found the employee","content": { "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } } } },"404": { "description": "Employee not found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } } } },"put": { "tags": [ "employee-controller" ],"summary": "Update an employee by id","operationId": "updateEmployee","parameters": [ { "name": "id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } } },"required": true },"responses": { "500": { "description": "Internal server error","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } },"200": { "description": "Updated employee","content": { "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } } } },"404": { "description": "Employee not found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } } } },"delete": { "tags": [ "employee-controller" ],"summary": "Delete employee","operationId": "deleteEmployee","parameters": [ { "name": "id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "200": { "description": "Confirmation message","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } },"500": { "description": "Internal server error","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } },"404": { "description": "Employee not found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } } } } },"/api/employees": { "get": { "tags": [ "employee-controller" ],"summary": "Get all employees","operationId": "showAllEmployees","responses": { "500": { "description": "Internal server error","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } },"200": { "description": "Array of employees","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/Employee" } } } } } } },"post": { "tags": [ "employee-controller" ],"summary": "Create employee","operationId": "addEmployee","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } } },"required": true },"responses": { "500": { "description": "Internal server error","content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageDTO" } } } },"200": { "description": "Created employee","content": { "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } } } } } } } },"components": { "schemas": { "Employee": { "required": [ "departmentId","firstName","jobTitle","lastName" ],"type": "object","properties": { "id": { "type": "integer","format": "int64" },"firstName": { "type": "string" },"lastName": { "type": "string" },"departmentId": { "type": "integer","format": "int32" },"jobTitle": { "type": "string" },"gender": { "type": "string","enum": [ "MALE","FEMALE" ] },"dateOfBirth": { "type": "string","format": "date-time" } } },"MessageDTO": { "type": "object","properties": { "message": { "type": "string" } } } } } }